requirements.txt
pdfminer.six==20221105
sentence-transformers==2.2.2
faiss-cpu==1.7.4
scikit-learn==1.3.0
numpy==1.26.4
streamlit==1.35.0
